+++ src/validator/tokenizer.h
@@ -15,7 +15,7 @@ enum tokenType
 int32_t yylex();
 int32_t yycurline();
 int32_t yycurpos();
-size_t yytoklen();
+uint32_t yytoklen();
 char* yytoktext();
 char* yytoktextcopy();
